+++ modules/mauth-authenticator/src/main/scala/com/mdsol/mauth/scaladsl/Authenticator.scala
@@ -17,4 +17,11 @@ trait Authenticator {
     * @return True or false indicating if the request is valid or not with respect to mAuth.
     */
   def authenticate(mAuthRequest: MAuthRequest)(implicit ex: ExecutionContext, requestValidationTimeout: Duration): Future[Boolean]
+
+  /**
+    * check if mauth v2 only authenticate is enabled or not
+    * @return True or false identifying if mauth v2 only is enabled or not.
+    */
+  def isV2OnlyAuthenticate: Boolean
+
 }