+++ modules/mauth-authenticator/src/main/scala/com/mdsol/mauth/scaladsl/RequestAuthenticator.scala
@@ -1,9 +1,11 @@
 package com.mdsol.mauth.scaladsl
 
 import java.nio.charset.StandardCharsets
+import java.security.PublicKey
 import java.util
 
 import com.mdsol.mauth.MAuthRequest
+import com.mdsol.mauth.MAuthVersion
 import com.mdsol.mauth.exception.MAuthValidationException
 import com.mdsol.mauth.scaladsl.utils.ClientPublicKeyProvider
 import com.mdsol.mauth.util.{EpochTimeProvider, MAuthSignatureHelper}
@@ -12,9 +14,13 @@ import org.slf4j.LoggerFactory
 import scala.concurrent.duration.Duration
 import scala.concurrent.{ExecutionContext, Future, Promise}
 
-class RequestAuthenticator(publicKeyProvider: ClientPublicKeyProvider, epochTimeProvider: EpochTimeProvider) extends Authenticator {
+class RequestAuthenticator(publicKeyProvider: ClientPublicKeyProvider, epochTimeProvider: EpochTimeProvider, v2OnlyAuthenticate: Boolean)
+    extends Authenticator {
 
   private val logger = LoggerFactory.getLogger(classOf[RequestAuthenticator])
+
+  def this(publicKeyProvider: ClientPublicKeyProvider, epochTimeProvider: EpochTimeProvider) =
+    this(publicKeyProvider, epochTimeProvider, false)
 
   /**
     * Performs the validation of an incoming HTTP request.
@@ -31,6 +37,10 @@ class RequestAuthenticator(publicKeyProvider: ClientPublicKeyProvider, epochTime
       val message = s"MAuth request validation failed because of timeout $requestValidationTimeout"
       logger.error(message)
       promise.failure(new MAuthValidationException(message))
+    } else if (!validateMauthVersion(mAuthRequest, v2OnlyAuthenticate)) {
+      val message = "The service requires mAuth v2 authentication headers."
+      logger.error(message)
+      promise.failure(new MAuthValidationException(message))
     } else {
       promise.completeWith(
         publicKeyProvider.getPublicKey(mAuthRequest.getAppUUID).map {
@@ -39,29 +49,72 @@ class RequestAuthenticator(publicKeyProvider: ClientPublicKeyProvider, epochTime
             false
           case Some(clientPublicKey) =>
             // Decrypt the signature with public key from requesting application.
-            val decryptedSignature = MAuthSignatureHelper.decryptSignature(clientPublicKey, mAuthRequest.getRequestSignature)
-
-            // Recreate the plain text signature, based on the incoming request parameters, and hash it.
-            val unencryptedRequestString =
-              MAuthSignatureHelper.generateUnencryptedSignature(
-                mAuthRequest.getAppUUID,
-                mAuthRequest.getHttpMethod,
-                mAuthRequest.getResourcePath,
-                new String(mAuthRequest.getMessagePayload, StandardCharsets.UTF_8),
-                String.valueOf(mAuthRequest.getRequestTime)
-              )
-            val messageDigest_bytes = MAuthSignatureHelper.getHexEncodedDigestedString(unencryptedRequestString).getBytes(StandardCharsets.UTF_8)
-
-            // Compare the decrypted signature and the recreated signature hashes.
-            // If both match, the request was signed by the requesting application and is valid.
-            util.Arrays.equals(messageDigest_bytes, decryptedSignature)
+            mAuthRequest.getMauthVersion match {
+              case MAuthVersion.MWS =>
+                validateSignatureV1(mAuthRequest, clientPublicKey)
+              case MAuthVersion.MWSV2 =>
+                validateSignatureV2(mAuthRequest, clientPublicKey)
+            }
         }
       )
     }
     promise.future
   }
 
+  /**
+    * check if mauth v2 only authenticate is enabled or not
+    * @return True or false identifying if v2 only authenticate is enabled or not.
+    */
+  override def isV2OnlyAuthenticate() = v2OnlyAuthenticate
+
   // Check epoch time is not older than specified interval.
   protected def validateTime(requestTime: Long)(requestValidationTimeout: Duration): Boolean =
     (epochTimeProvider.inSeconds - requestTime) < requestValidationTimeout.toSeconds
+
+  // Check V2 header if only V2 is required
+  protected def validateMauthVersion(mAuthRequest: MAuthRequest, v2OnlyAuthenticate: Boolean): Boolean =
+    !v2OnlyAuthenticate || mAuthRequest.getMauthVersion == MAuthVersion.MWSV2
+
+  // check signature for V1
+  private def validateSignatureV1(mAuthRequest: MAuthRequest, clientPublicKey: PublicKey): Boolean = {
+    val decryptedSignature = MAuthSignatureHelper.decryptSignature(clientPublicKey, mAuthRequest.getRequestSignature)
+    // Recreate the plain text signature, based on the incoming request parameters, and hash it.
+    val unencryptedRequestString = MAuthSignatureHelper.generateUnencryptedSignature(
+      mAuthRequest.getAppUUID,
+      mAuthRequest.getHttpMethod,
+      mAuthRequest.getResourcePath,
+      new String(mAuthRequest.getMessagePayload, StandardCharsets.UTF_8),
+      String.valueOf(mAuthRequest.getRequestTime)
+    )
+    val messageDigest_bytes = MAuthSignatureHelper.getHexEncodedDigestedString(unencryptedRequestString).getBytes(StandardCharsets.UTF_8)
+
+    // Compare the decrypted signature and the recreated signature hashes.
+    // If both match, the request was signed by the requesting application and is valid.
+    util.Arrays.equals(messageDigest_bytes, decryptedSignature)
+  }
+
+  // check signature for V2
+  private def validateSignatureV2(mAuthRequest: MAuthRequest, clientPublicKey: PublicKey): Boolean = {
+    // Recreate the plain text signature, based on the incoming request parameters, and hash it.
+    val unencryptedRequestString = MAuthSignatureHelper.generateStringToSignV2(
+      mAuthRequest.getAppUUID,
+      mAuthRequest.getHttpMethod,
+      mAuthRequest.getResourcePath,
+      mAuthRequest.getQueryParameters,
+      mAuthRequest.getMessagePayload,
+      String.valueOf(mAuthRequest.getRequestTime)
+    )
+
+    // Compare the decrypted signature and the recreated signature hashes.
+    try {
+      MAuthSignatureHelper.verifyRSA(unencryptedRequestString, mAuthRequest.getRequestSignature, clientPublicKey)
+    } catch {
+      case ex: Exception =>
+        val message = "MAuth request validation failed because of " + ex.getMessage
+        logger.error(message)
+        return false
+    }
+
+  }
+
 }