+++ modules/mauth-signer-akka-http/src/main/scala/com/mdsol/mauth/http/Implicits.scala
@@ -18,11 +18,19 @@ object Implicits {
     }
     val contentType: Option[String] = extractContentTypeFromHeaders(sr.req.headers)
     val headersWithoutContentType: Map[String, String] = removeContentTypeFromHeaders(sr.req.headers)
-
-    HttpRequest(method = sr.req.httpMethod, uri = Uri(sr.req.uri.toString), entity = getHttpEntity(contentType, entityBody))
-      .withHeaders(
-        mapToHeaderSequence(headersWithoutContentType) ++: scala.collection.immutable.Seq(`X-MWS-Authentication`(sr.authHeader), `X-MWS-Time`(sr.timeHeader))
-      )
+    if (sr.mauthHeaders.nonEmpty) {
+      val headers: scala.collection.immutable.Seq[HttpHeader] = sr.mauthHeaders.map { case (k, v) => RawHeader(k, v) }.to[scala.collection.immutable.Seq]
+      HttpRequest(method = sr.req.httpMethod, uri = Uri(sr.req.uri.toString), entity = getHttpEntity(contentType, entityBody))
+        .withHeaders(
+          mapToHeaderSequence(headersWithoutContentType) ++: headers
+        )
+    } else {
+      // Mauth V1 only for binary compatibility
+      val headers = mapToHeaderSequence(headersWithoutContentType) ++: scala.collection.immutable
+        .Seq(`X-MWS-Authentication`(sr.authHeader), `X-MWS-Time`(sr.timeHeader))
+      HttpRequest(method = sr.req.httpMethod, uri = Uri(sr.req.uri.toString), entity = getHttpEntity(contentType, entityBody))
+        .withHeaders(headers)
+    }
   }
 
   implicit def fromMaybeSignedRequestToMaybeHttpRequest(maybeSignedRequest: Option[SignedRequest]): Option[HttpRequest] =