+++ modules/mauth-signer/src/main/java/com/mdsol/mauth/DefaultSigner.java
@@ -11,6 +11,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.nio.charset.StandardCharsets;
 import java.security.PrivateKey;
 import java.security.Security;
 import java.util.HashMap;
@@ -25,7 +26,7 @@ public class DefaultSigner implements Signer {
   private final UUID appUUID;
   private final PrivateKey privateKey;
   private final EpochTimeProvider epochTimeProvider;
-  private boolean disableV1 = false;
+  private boolean v2OnlySignRequests = false;
 
   static {
     Security.addProvider(new BouncyCastleProvider());
@@ -33,7 +34,7 @@ public class DefaultSigner implements Signer {
 
   public DefaultSigner(SignerConfiguration configuration) {
     this(configuration.getAppUUID(), getPrivateKeyFromString(configuration.getPrivateKey()),
-        new CurrentEpochTimeProvider(), configuration.isDisableV1());
+        new CurrentEpochTimeProvider(), configuration.isV2OnlySignRequests());
   }
 
   public DefaultSigner(UUID appUUID, String privateKey) {
@@ -52,15 +53,15 @@ public class DefaultSigner implements Signer {
     this(appUUID, privateKey, epochTimeProvider, false);
   }
 
-  public DefaultSigner(UUID appUUID, String privateKey, EpochTimeProvider epochTimeProvider, boolean disableV1) {
-    this(appUUID, getPrivateKeyFromString(privateKey), epochTimeProvider, disableV1);
+  public DefaultSigner(UUID appUUID, String privateKey, EpochTimeProvider epochTimeProvider, boolean v2OnlySignRequests) {
+    this(appUUID, getPrivateKeyFromString(privateKey), epochTimeProvider, v2OnlySignRequests);
   }
 
-  public DefaultSigner(UUID appUUID, PrivateKey privateKey, EpochTimeProvider epochTimeProvider, boolean disableV1) {
+  public DefaultSigner(UUID appUUID, PrivateKey privateKey, EpochTimeProvider epochTimeProvider, boolean v2OnlySignRequests) {
     this.appUUID = appUUID;
     this.privateKey = privateKey;
     this.epochTimeProvider = epochTimeProvider;
-    this.disableV1 = disableV1;
+    this.v2OnlySignRequests = v2OnlySignRequests;
   }
 
   @Override
@@ -78,10 +79,10 @@ public class DefaultSigner implements Signer {
 
   @Override
   public Map<String, String> generateRequestHeaders(String httpVerb,
-      String requestPath, String requestPayload, String queryParameters) throws MAuthSigningException {
+      String requestPath, byte[] requestPayload, String queryParameters) throws MAuthSigningException {
 
     if (null == requestPayload) {
-      requestPayload = "";
+      requestPayload = "".getBytes(StandardCharsets.UTF_8);
     }
     if (null == queryParameters) {
       queryParameters = "";
@@ -91,9 +92,9 @@ public class DefaultSigner implements Signer {
     long currentTime = epochTimeProvider.inSeconds();
 
     HashMap<String, String> headers = new HashMap<>();
-    if (!disableV1) {
-      // Add v1 headers if Mauth V1 is not disabled
-      Map<String, String> headersV1 = generateRequestHeadersV1(httpVerb, requestPath, requestPayload, currentTime);
+    if (!v2OnlySignRequests) {
+      // Add v1 headers if not V2 only sign requests
+      Map<String, String> headersV1 = generateRequestHeadersV1(httpVerb, requestPath, new String(requestPayload, StandardCharsets.UTF_8), currentTime);
       if (!headersV1.isEmpty()) {
         headers.putAll(headersV1);
       }
@@ -131,10 +132,9 @@ public class DefaultSigner implements Signer {
   }
 
   private Map<String, String> generateRequestHeadersV2(String httpVerb, String requestPath, String queryParameters,
-      String requestPayload, long currentTime) throws MAuthSigningException {
-    MAuthVersion mauthVersion = MAuthVersion.MWSV2;
-    String stringTosign = MAuthSignatureHelper.generateStringToSign(
-        appUUID, httpVerb, requestPath, queryParameters, requestPayload, String.valueOf(currentTime), mauthVersion);
+      byte[] requestPayload, long currentTime) throws MAuthSigningException {
+     String stringTosign = MAuthSignatureHelper.generateStringToSignV2(
+        appUUID, httpVerb, requestPath, queryParameters, requestPayload, String.valueOf(currentTime));
 
     String encryptedSignature;
     try {
@@ -147,7 +147,7 @@ public class DefaultSigner implements Signer {
     HashMap<String, String> headers = new HashMap<>();
     headers.put(
         MAuthRequest.MCC_AUTHENTICATION_HEADER_NAME,
-        MAuthHeadersHelper.createAuthenticationHeaderValue(appUUID, encryptedSignature, mauthVersion.getValue())
+        MAuthHeadersHelper.createAuthenticationHeaderValue(appUUID, encryptedSignature, MAuthVersion.MWSV2.getValue())
     );
     headers.put(MAuthRequest.MCC_TIME_HEADER_NAME, MAuthHeadersHelper.createTimeHeaderValue(currentTime));