+++ modules/mauth-signer/src/main/java/com/mdsol/mauth/SignerConfiguration.java
@@ -9,16 +9,17 @@ public class SignerConfiguration implements MAuthConfiguration {
   public static final String MAUTH_SECTION_HEADER = "mauth";
   public static final String APP_UUID_PATH = APP_SECTION_HEADER + ".uuid";
   public static final String APP_PRIVATE_KEY_PATH = APP_SECTION_HEADER + ".private_key";
-  public static final String DISABLE_MAUTH_V1 = MAUTH_SECTION_HEADER + ".disable_v1";
+  public static final String V2_ONLY_SIGN_REQUESTS = MAUTH_SECTION_HEADER + ".v2_only_sign_requests";
+
 
   private final UUID appUUID;
   private final transient String privateKey;
-  private boolean disableV1 = false;
+  private boolean v2OnlySignRequests = false;
 
   public SignerConfiguration(Config config) {
     this( UUID.fromString(config.getString(APP_UUID_PATH)),
         config.getString(APP_PRIVATE_KEY_PATH),
-        config.hasPath(DISABLE_MAUTH_V1) || config.isEmpty() ? config.getBoolean(DISABLE_MAUTH_V1) : false);
+        config.hasPath(V2_ONLY_SIGN_REQUESTS) || config.isEmpty() ? config.getBoolean(V2_ONLY_SIGN_REQUESTS) : false);
   }
 
   public SignerConfiguration(UUID appUUID, String privateKey) {
@@ -29,13 +30,13 @@ public class SignerConfiguration implements MAuthConfiguration {
     this.privateKey = privateKey;
   }
 
-  public SignerConfiguration(UUID appUUID, String privateKey, boolean disableV1) {
+  public SignerConfiguration(UUID appUUID, String privateKey, boolean v2OnlySignRequests) {
     validateNotNull(appUUID, "Application UUID");
     validateNotBlank(privateKey, "Application Private key");
 
     this.appUUID = appUUID;
     this.privateKey = privateKey;
-    this.disableV1 = disableV1;
+    this.v2OnlySignRequests = v2OnlySignRequests;
   }
 
   public UUID getAppUUID() {
@@ -46,7 +47,7 @@ public class SignerConfiguration implements MAuthConfiguration {
     return privateKey;
   }
 
-  public boolean isDisableV1() {
-    return disableV1;
+  public boolean isV2OnlySignRequests() {
+    return v2OnlySignRequests;
   }
 }